+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.2.2 (Unreleased)
 
 ### Improvments
-- [databricks] - Boost Databricks engineering runtime to 0.5.2  
+- [databricks] - Boost Databricks engineering runtime to 0.6.0  
 - [databricks/provider] - Bump the version to 0.4.1. Also lock the version to the exact semver.
 - [docker_image/iac_runtime] - Create a new Dockerfile. Use the same IaC runtime for the platform deployment and for the devcontainers.
 - [platform/source] - Publish the source of the platform in the Docker image (/platform/src). Make the Docker image `private` in the dockerhub.